requirements.txt
Babel==2.3.4
defusedxml==0.4.1
Django==1.10.5
django-allauth==0.30.0
django-articles==2.4.1
django-articles-addons==0.2.4
django-autoslug==1.9.3
django-bootstrap-breadcrumbs==0.8.2
django-bootstrap-themes==3.3.6
django-chartit==0.2.7
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-phonenumber-field==1.1.0
django-smtp-ssl==1.0
django-tables2==1.2.9
markdown2==2.3.2
oauthlib==2.0.1
olefile==0.44
phonenumberslite==8.0.0
Pillow==4.0.0
prettytable==0.7.2
python3-openid==3.0.10
pytz==2016.10
reportlab==3.3.0
requests==2.12.5
requests-oauthlib==0.7.0
six==1.10.0
urllib3==1.20
